pyproject.toml
[project]
name = "rag-fastapi-poc"
version = "0.1.0"
description = "RAG POC with FastAPI, Ollama, and local embeddings"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "transformers>=4.46.0",
    "torch>=2.5.0",
    "sentence-transformers>=3.3.0",
    "faiss-cpu>=1.9.0",
    "httpx>=0.27.0",
    "pypdf>=5.1.0",
    "numpy>=2.1.0",
    "python-multipart>=0.0.12",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
]
